Move auth and company to ContextProvider pattern
Also move code out of App.tsx to the context components.
