+++ src/index.tsx
@@ -2,11 +2,5 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import App from './App';
-import { AuthProvider } from './store/contexts/auth';
 
-ReactDOM.render(
-  <AuthProvider>
-    <App />
-  </AuthProvider>,
-  document.getElementById('root')
-);
+ReactDOM.render(<App />, document.getElementById('root'));
